ci: add a step for running tests
