+++ helpers/utils.js
@@ -39,3 +39,22 @@ export const getDaysDifference = (startDate, endDate) => {
   const end = new Date(endDate);
   return (end.getTime() - start.getTime()) / (1000 * 60 * 60 * 24);
 };
+
+export const getTimestampsForBatch = (currentIndex, maxDays) => {
+  const BATCH_SIZE = 10; // 10 days
+  let batchStart = currentIndex == 0 ? BATCH_SIZE : currentIndex + BATCH_SIZE;
+
+  const difference = Math.abs(maxDays - currentIndex);
+  let batchEnd = difference < BATCH_SIZE ? difference : 10;
+
+  const startTime = new Date();
+  startTime.setDate(startTime.getDate() - batchStart);
+
+  const endTime = new Date(startTime);
+  endTime.setDate(startTime.getDate() + batchEnd);
+
+  return {
+    startTime: startTime.toISOString(),
+    endTime: endTime.toISOString(),
+  };
+};
